+++ examples/ref_column_models.py
@@ -1,3 +1,3 @@
 from open_alchemy import init_yaml
 
-init_yaml("ref-column-example-spec.yml")
+init_yaml("ref-column-example-spec.yml", models_filename="ref_column_models_auto.py")