+++ examples/ref_model_models.py
@@ -1,3 +1,3 @@
 from open_alchemy import init_yaml
 
-init_yaml("ref-model-example-spec.yml")
+init_yaml("ref-model-example-spec.yml", models_filename="ref_model_models_auto.py")