+++ examples/relationship_one_to_many_models.py
@@ -1,3 +1,6 @@
 from open_alchemy import init_yaml
 
-init_yaml("relationship-one-to-many-example-spec.yml")
+init_yaml(
+    "relationship-one-to-many-example-spec.yml",
+    models_filename="relationship_one_to_many_models_auto.py",
+)