+++ open_alchemy/helpers/define_all.py
@@ -1,5 +1,6 @@
 """Define all the models with x-tablename properties."""
 
+from open_alchemy import helpers
 from open_alchemy import types
 
 
@@ -13,6 +14,10 @@ def define_all(*, model_factory: types.ModelFactory, schemas: types.Schemas) -> 
 
     """
     for name, schema in schemas.items():
-        if schema.get("x-tablename") is None:
+        # Skip models that do not have a x-tablename defined
+        if helpers.peek.tablename(schema=schema, schemas=schemas) is None:
+            continue
+        # Skip models that just reference other models
+        if schema.get("$ref") is not None:
             continue
         model_factory(name=name)