+++ tests/open_alchemy/column_factory/test_column.py
@@ -12,43 +12,6 @@ from open_alchemy.column_factory import column
 
 
 @pytest.mark.parametrize(
-    "required, nullable, expected_result",
-    [
-        (None, None, True),
-        (None, False, False),
-        (None, True, True),
-        (False, None, True),
-        (False, False, False),
-        (False, True, True),
-        (True, None, False),
-        (True, False, False),
-        (True, True, True),
-    ],
-    ids=[
-        "required not given nullable not given",
-        "required not given nullable reset",
-        "required not given nullable set",
-        "required reset nullable not given",
-        "required reset nullable reset",
-        "required reset nullable set",
-        "required set nullable not given",
-        "required set nullable reset",
-        "required set nullable set",
-    ],
-)
-@pytest.mark.column
-def test_calculate_nullable(required, nullable, expected_result):
-    """
-    GIVEN required, nullable and expected result
-    WHEN _calculate_nullable is called with nullable and required
-    THEN the expected result is returned.
-    """
-    result = column._calculate_nullable(nullable=nullable, required=required)
-
-    assert result == expected_result
-
-
-@pytest.mark.parametrize(
     "schema, expected_exception",
     [
         ({}, exceptions.TypeMissingError),