+++ tests/open_alchemy/helpers/test_define_all.py
@@ -13,6 +13,14 @@ from open_alchemy import helpers
         ({}, []),
         ({"Table": {}}, []),
         ({"Table": {"x-tablename": "table"}}, ["Table"]),
+        ({"Table": {"allOf": [{"x-tablename": "table"}]}}, ["Table"]),
+        (
+            {
+                "Table": {"x-tablename": "table"},
+                "RefTable": {"$ref": "#/components/schemas/Table"},
+            },
+            ["Table"],
+        ),
         ({"Table1": {}, "Table2": {}}, []),
         ({"Table1": {"x-tablename": "table1"}, "Table2": {}}, ["Table1"]),
         ({"Table1": {}, "Table2": {"x-tablename": "table2"}}, ["Table2"]),
@@ -25,6 +33,8 @@ from open_alchemy import helpers
         "empty,                          zero",
         "single no x-tablename,          zero",
         "single x-tablename,             one",
+        "allOf single x-tablename,       one",
+        "$ref single x-tablename,        one",
         "multiple no x-tablename,        multiple",
         "multiple one first x-tablename, multiple",
         "multiple one last x-tablename, multiple",