+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 - Add support for _date_
 - Move SQLAlchemy relationship construction behind facade
 - Move schema calculations into separate files
-- Refactor handling array references to reduce scope of individual tests and make them easier to understand.
+- Refactor handling array references to reduce scope of individual tests and make them easier to understand
+- Add optional parameter that can be used to generate a models file for IDE auto complete and type hinting
 
 ## Version 0.10.4 - 2019-12-18